fix(ci): Use --lib --tests for Clippy instead of --all-targets
The --all-targets flag tries to build the binary which requires
the frontend dist directory to exist. This causes a proc macro
panic: "The `frontendDist` configuration is set to `../dist`
but this path doesn't exist"

Changed to --lib --tests which only checks:
- Library code (where all our logic is)
- Test code

This is sufficient for our CI needs and doesn't require building
the full Tauri app (which the separate build-tauri job does).